backend/requirements.txt
# Core dependencies for disaster mapping module
numpy>=1.21.0
opencv-python>=4.5.0
scikit-image>=0.19.0
shapely>=1.8.0
folium>=0.12.0

# GeoTIFF support (optional but recommended)
rasterio>=1.2.0

# Firebase integration (optional)
firebase-admin>=5.0.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Utilities
Pillow>=9.0.0

